pyproject.toml
[project]
name = "cc-streamdeck"
version = "0.1.0"
description = "Stream Deck Mini interface for Claude Code permission prompts"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "streamdeck>=0.9.5",
    "Pillow>=10.0",
]

[project.scripts]
cc-streamdeck-daemon = "cc_streamdeck.daemon:main"
cc-streamdeck-hook = "cc_streamdeck.hook:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cc_streamdeck"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]
